+++ packages/baseten/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/baseten
 
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- 0c4822d: feat: `EmbeddingModelV3`
+- 1cad0ab: feat: add provider version to user-agent header
+- Updated dependencies [0c4822d]
+  - @ai-sdk/openai-compatible@1.1.0-beta.2
+  - @ai-sdk/provider@2.1.0-beta.1
+  - @ai-sdk/provider-utils@3.1.0-beta.2
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Updated dependencies [cbb1d35]
+  - @ai-sdk/provider-utils@3.1.0-beta.1
+  - @ai-sdk/openai-compatible@1.1.0-beta.1
+
 ## 1.0.0-beta.0
 
 ### Major Changes