+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/provider-utils
 
+## 3.1.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [0c4822d]
+  - @ai-sdk/provider@2.1.0-beta.1
+
+## 3.1.0-beta.1
+
+### Patch Changes
+
+- cbb1d35: Update for provider-util changeset after change in PR #8588
+
 ## 3.1.0-beta.0
 
 ### Minor Changes